pyproject.toml
[tool.poetry]
name = "blochbuster"
version = "1.0.0"
description = "BlochBuster is a nuclear magnetic resonance Bloch equation simulator that creates animated plots."
authors = ["Johan Berglund <johan.berglund@akademiska.se>"]
license = "GPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/bretglun/BlochBuster"

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
scipy = "^1.9.3"
matplotlib = "^3.7.3"
PyYAML = "^6.0.1"
ffmpeg-python = "^0.2.0"
static-ffmpeg = "^2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
